Add message loss protection to retryMonitor
Previously we had a small but non-zero risk of the retry monitor losing messages, as messages were hold only in the local process memory between being popped from the retry set and added back to the queue.

This change guarantees that will not happen by using atomic transactions with an optimistic locking strategy. It also has the added benefit of making the client code simpler :)

closes Use transactions when doing retries #13